Remove annoying console "error" - also don't use '0' as an error-defindex (that's a valid defindex)

+++ SteamTrade/Trade.cs
@@ -693,7 +693,7 @@ namespace SteamTrade
                 OtherPrivateInventory = new ForeignInventory(foreignInventory);
             }
 
-            ushort defindex = OtherPrivateInventory.GetDefIndex(asset.assetid);
+            int defindex = OtherPrivateInventory.GetDefIndex(asset.assetid);
             Schema.Item schemaItem = CurrentSchema.GetItem(defindex);
             return schemaItem;
         }
